refactor: Location 클래스의 패키지 경로 변경 및 관련 import 수정
