+++ tomcat/src/main/java/com/techcourse/handler/RegisterRequestHandler.java
@@ -6,9 +6,9 @@ import com.techcourse.http.common.ContentType;
 import com.techcourse.http.common.HttpMethod;
 import com.techcourse.http.common.HttpStatus;
 import com.techcourse.http.common.HttpVersion;
-import com.techcourse.http.common.Location;
 import com.techcourse.http.request.HttpRequest;
 import com.techcourse.http.response.HttpResponse;
+import com.techcourse.http.response.Location;
 import com.techcourse.model.User;
 import com.techcourse.util.FileUtil;
 import java.util.Map;
