+++ tomcat/src/main/java/com/techcourse/http/response/Location.java
@@ -1,4 +1,4 @@
-package com.techcourse.http.common;
+package com.techcourse.http.response;
 
 public class Location {
