+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -6,9 +6,9 @@ import com.techcourse.handler.RegisterRequestHandler;
 import com.techcourse.http.common.ContentType;
 import com.techcourse.http.common.HttpStatus;
 import com.techcourse.http.common.HttpVersion;
-import com.techcourse.http.common.Location;
 import com.techcourse.http.request.HttpRequest;
 import com.techcourse.http.response.HttpResponse;
+import com.techcourse.http.response.Location;
 import com.techcourse.util.FileUtil;
 import java.io.BufferedReader;
 import java.io.IOException;
